requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.12
langchain-google-genai>=0.0.11
langchain-groq>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
typing-extensions>=4.7.0

# CLI and formatting
rich>=13.0.0
python-dateutil>=2.8.2

# Web and API
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.2
python-multipart>=0.0.6
python-socketio>=5.9.0
websockets>=11.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Frontend dependencies
python-frontend>=0.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
